rename snapshot module
